refactor(game-engine): extract choice handlers (S27.8.7)
Slice 7 de S27.8 (refactor monolith `actions.ts`).

Extraction (`actions/choice-handlers.ts`)
- `handleBlockChoose` : choix du resultat de des de bloc apres
  un blocage qui a roule plusieurs des.
- `handlePushChoose` : choix de la direction de poussee parmi
  celles disponibles (Fend, Frenzy, chain push).
- `handleFollowUpChoose` : confirmation du suivi (follow-up)
  apres une poussee classique.
- `handleRerollChoose` : decision sur l'utilisation d'une relance
  d'equipe pour un jet rate (dodge / GFI / pickup), avec test
  Solitaire (Loner) le cas echeant.

Helpers internes deplaces avec `handleRerollChoose` (consommes
uniquement par lui)
- `getLonerThreshold(player)`
- `consumeTeamReroll(state, team)`

Imports orphelins retires de `actions.ts`
- `applyChainPush` (mechanics/blocking)
- `isFendActiveForFollowUp` (mechanics/fend)
- `hasFrenzy` (mechanics/frenzy)
- `performPickupRollWithNotification` (utils/dice-notifications)

Aucune dependance interne au dispatcher : tout passe par les
modules deja extraits (`failure-helpers`, `ball-pickup`) +
`core/game-state` + `mechanics/*` + `skills/*`.

Resultat
- `actions.ts` 1878 -> 1517 lignes (-361 nettes ; cumul -1189 vs
  origine 2706 = ~43.9%).
- `choice-handlers.ts` ajoute a la liste d'exclusion coverage.
- 4888 tests game-engine passent. Coverage 25.33% maintenue.

Reste pour atteindre la cible <= 600 lignes :
- handleBlock + handleMultiBlock + handleDumpOffChoose +
  resolveFrenzyBlock + resolveMultipleBlock (block subsystem)
- handleBlitz
- handleLeap / handleDodge / handleMove (couplage getLegalMoves)
