+++ packages/game-engine/vitest.config.ts
@@ -39,6 +39,7 @@ export default defineConfig({
         'src/actions/failure-helpers.ts',
         'src/actions/ball-pickup.ts',
         'src/actions/move-handlers.ts',
+        'src/actions/choice-handlers.ts',
       ],
       // S25.4 — Seuils initiaux conservateurs. v8 reporte 0% lines sur
       // certains fichiers TS du moteur a cause d'un quirk source-map ;
